* remotes/origin/master:
  IMPRO-1917: Boolean output are now int8 from these plugins/CLIs: (#1362)
  IMPRO-1815: reject float64 data in combine plugin (#1363)

+++ improver/psychrometric_calculations/precip_phase_probability.py
@@ -177,7 +177,7 @@ class PrecipPhaseProbability(BasePlugin):
             self.comparator(self.orography_cube.data, processed_falling_level.data),
             1,
             0,
-        ).astype("float32")
+        ).astype(np.int8)
         mandatory_attributes = generate_mandatory_attributes([self.falling_level_cube])
 
         cube = create_new_diagnostic_cube(
